requirements.txt
accelerate==1.0.1
bitsandbytes==0.44.1
langchain-community==0.3.2
langchain-core==0.3.12
langchain-huggingface==0.1.0.dev1
langchain-qdrant==0.2.0.dev1
langchain-text-splitters==0.3.0
torch~=2.4.0
transformers~=4.43.1
xformers==0.0.25.post1
qdrant-client==1.12.0
sentence-transformers==3.2.0
pm4py==2.7.11.8
python-dotenv==1.0.1
sentencepiece==0.2.0
pandas~=2.2.2
numpy~=1.26.4
matplotlib~=3.9.1
scikit-learn~=1.5.0
datasets~=2.19.1